market_manipulation/requirements.txt
absl-py==0.10.0
appdirs==1.4.4
appnope==0.1.2
astunparse==1.6.3
backcall==0.2.0
bsuite==0.3.5
cachetools==4.1.1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
distlib==0.3.1
dm-env==1.4
dm-sonnet==2.0.0
dm-tree==0.1.6
filelock==3.0.12
flake8==3.9.1
frozendict==1.2
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
identify==2.2.4
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.2
-e git+git@github.com:MaxOSmith/market_manipulation.git@fdd51513ec6a7c4b0c0ae3476d92931697a11bb0#egg=market_manipulation
matplotlib==3.4.1
mccabe==0.6.1
mizani==0.7.3
networkx==2.5.1
nodeenv==1.6.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
palettable==3.3.0
pandas==1.2.4
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotnine==0.8.0
pre-commit==2.12.1
prompt-toolkit==3.0.18
protobuf==3.13.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyglet==1.5.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
statsmodels==0.12.2
tabulate==0.8.9
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.25.10
virtualenv==20.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
